requirements.txt
flask
azure-cognitiveservices-speech
python-dotenv